Add button for in app browser popup to be able to open the site inside the browser
